requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
transformers==4.36.2
torch==2.1.2
torchaudio==2.1.2
sentence-transformers==2.3.1
faiss-cpu==1.7.4
pydub==0.25.1
numpy==1.24.3
accelerate==0.25.0
